Update class_wdeb_admin_form_renderer.php
To avoid PHP Fatal error on line 12.  "Uncaught Error: Cannot access offset of type string on string on line 12"
The error occurs,  because it tries to access an offset on a string, which is not allowed in PHP. In this case, the issue is with the $opt variable. If the option doesn't exist or is not an array, it will return a string instead of an array.

To fix this error, we ensure that the $opt variable is an array before trying to access its offsets.